get_repos_function/requirements.txt
PyYAML==5.3
requests==2.23.0
google-cloud-secret-manager==0.2.0
google-cloud-pubsub==1.3.1
google-api-core==1.16.0